+++ CANIM_class/CCamera.h
@@ -1,7 +1,7 @@
 #ifndef _CCAMERA_H_
     #define _CCAMERA_H_
 #include <SDL.h>
-#include "Define.h"
+#include "../Define.h"
 enum{
     TARGET_MODE_NORMAL=0,
     TARGET_MODE_CENTER
@@ -15,8 +15,8 @@ class CCamera{
         int X;
         int Y;
         //Координаты цели камеры
-        int* TargetX;
-        int* TargetY;
+        float* TargetX;
+        float* TargetY;
     public:
         //Режим нацеливания камеры
         int TargetMode;
@@ -33,6 +33,6 @@ class CCamera{
     public:
         //Установка координат камеры и цели
         void SetPos(int X, int Y);
-        void SetTarget(int* X, int* Y);
+        void SetTarget(float* X, float* Y);
 };
 #endif
